feat: Enhance documentation and examples; update congestion control algorithms and connection stats structure - Updated DOCUMENTATION.md to reflect new frame types and structures. - Improved README.md with comprehensive examples for reliable streams, datagrams, and zero-copy transmission. - Modified ZtConfig to use u64 for initial_max_data for better flow control. - Refactored key derivation functions to enhance security and zeroize sensitive data. - Updated connection and stream handling to support new congestion control algorithms and improved telemetry. - Enhanced error handling and logging in event loop and packet sender. - Improved stream buffer management for efficiency and performance. - Added notes in integration tests for reliable execution under extreme conditions.

+++ src/config.rs
@@ -35,7 +35,7 @@ pub struct ZtConfig {
     /// Maximum stream receive window after auto-tuning in bytes. Default: 16 MB.
     pub max_stream_window: u64,
     /// Initial connection-level flow control window in bytes. Default: 1 MB.
-    pub initial_max_data: u32,
+    pub initial_max_data: u64,
     /// Maximum total stream buffer memory per connection. Default: 64 MB.
     pub max_connection_buffer: usize,
